fix: do not marshal NaN and Inf float values (#34)
* fix: do not marshal NaN and Inf float values

NaN and Inf are not valid float values in json
the Go stdlib does not allow to marshal them and return an error.

fastjson does not return an error so the only option (besides panic)
is to treat them as 0.
Users of this library should validate they are properly handling
NaN/Inf float values if necessary

* refactor: move check up one level and return an error

when using Float method directly the user is responsible to
check for invalid float values

add test to ensure marshaling invalid floats return an error